Add router-selective display on RoutesPage.
In case of multiple peering routers of one customer, their
announcements are aggregated in a single per-customer table.
RoutesPage used to display contents of that table irrespectively
of the current router selected in the UI.
Now, when selecting a router in the UI RoutesPage is initialized
with the nextHop IP of that router to display only routes
(filtered, accepted and noexport) from that router.