Refactor route counts on protocols page.
Extend API to include filtered routes on pipe protocol. Required, hence in
case of multiple peering routers of one IXP participant we can only assume that
all routers advertise the same prefixes. Obviously, this assumption does not
hold true if the amount of advertised prefixes differs between those routers.
In essence we estimate the number of routers and correct the amount of accepted
prefixes and those filtered on the participant pipe towards the master table.