Fix routes deduplication for remote-triggered blackholing.
In case of remote-triggered blackholing filtered routes may be present
that have a nextHop different from the IP of the neighbour the route
has been learnt from. These would not appear in on the routes page.
With this patch, routes ar displayed if either
gateway matches nextHop or the learnt_from address.