Get the Network/AF from the NLRI in gobgp
This moves from parsing the prefix string to determine the Address
Family and changes it to inspect the NLRI itself. 

This is a minor change bug internally at Coreweave we are looking at
parsing other AFI/SAFI's than just unicast-ipv4 and unicast-ipv6 and
this is a good start to parsing the update itself.